Tooltip cleanup: dedupe entries, clearer date and time formats
- Cumulative chart: collapse separate Area+Line into single Area
  (Recharts emits one tooltip entry per series — having two with the
  same dataKey gave the doubled 'Cumulative : 0' lines).
- Tooltip values now read '3 calls' / '0 calls' instead of
  'Cumulative : 0'. Time label is '3:00 PM' not '3p' (short form
  reserved for the cramped x-axis ticks).
- Bar tooltip: 'Wed, May 6' instead of 'Wed · 5/6'.

+++ src/app/(dash)/page.tsx
@@ -367,7 +367,7 @@ function labelDate(localNow: Date, isoDay: number, todayIso: number): string {
   const offset = isoDay - todayIso;
   const d = new Date(localNow);
   d.setDate(d.getDate() + offset);
-  return d.toLocaleDateString("en-US", { month: "numeric", day: "numeric" });
+  return d.toLocaleDateString("en-US", { month: "short", day: "numeric" });
 }
 
 function workdaysSinceMonday(localNow: Date, workdays: readonly number[]): number {
